* Improvements in error handling. * Fix for includes. * Fix for types.
git-svn-id: https://studio2000.xyz:8443/svn/TPLTM/Lab1/Calc@249 bea4479c-5f86-c742-a837-c53f34f90b20

+++ Calc/token.h
@@ -8,10 +8,10 @@ struct Token
 	/*
 	 * Type using as a type for calculation.
 	 */
-	using Value = long double;
+	using Value = double;
 	using ParamCount = unsigned short;
 
-	enum class Type : unsigned int
+	enum class Type : unsigned short
 	{
 		END = 0,
 		ADD,
